Finished main algorithim functions and tested 1 operator
In the second class, Mahdi and I implemented our algorithm's main logic and tested it with one math operation (square). We ensured to finish the main logic together to divide the remaining operations between us with full understanding. We used the eval function and gave it access to chosen math functions, we also added some error cases that we want to prevent in the future.