Update Calculator.py
Mahdi and I used pair programming so he can add all of the functions that he's responsible for. after that, we looked at a science calculation to add things we missed (point,csc, sec, cot). Lastly, we realized the logical error when the user uses trig function with degrees as input, and fixed that as well.